Remove ember-dependency-lint
This linting test is not part of the normal Ember addon temp and is
blocking builds. I think it would be better to do this linting on the
top level project (editor, or frontend-gn) rather than only on this
specific plugin, or maybe on every plugin at once. It's just too
inconsistent for now.